refactor: API 요청 및 응답값으로 DTO 사용
