+++ src/main/java/roomescape/ReservationController.java
@@ -11,6 +11,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import roomescape.dto.ReservationFindAllResponse;
+import roomescape.dto.ReservationFindResponse;
+import roomescape.dto.ReservationSaveRequest;
 
 @RequestMapping("/reservations")
 @Controller
@@ -20,15 +23,19 @@ public class ReservationController {
     private final List<Reservation> reservations = new ArrayList<>();
 
     @GetMapping
-    public ResponseEntity<List<Reservation>> findAllReservation() {
-        return ResponseEntity.ok().body(reservations);
+    public ResponseEntity<List<ReservationFindAllResponse>> findAllReservation() {
+        List<ReservationFindAllResponse> response = reservations.stream()
+                .map(ReservationFindAllResponse::from)
+                .toList();
+        return ResponseEntity.ok().body(response);
     }
 
     @PostMapping
-    public ResponseEntity<Reservation> saveReservation(@RequestBody ReservationCreateRequest request) {
+    public ResponseEntity<ReservationFindResponse> saveReservation(@RequestBody ReservationSaveRequest request) {
         Reservation reservation = request.toEntity(counter.incrementAndGet());
         reservations.add(reservation);
-        return ResponseEntity.ok(reservation);
+        ReservationFindResponse response = ReservationFindResponse.from(reservation);
+        return ResponseEntity.ok(response);
     }
 
     @DeleteMapping("/{reservation_id}")
